requirements.txt
flask==3.0.0
python-dotenv==1.0.0
googletrans==3.1.0a0
flask-socketio==5.3.6
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
